componentize todos
